Use standard require + require_relative
The require_lib stuff is not finding the right paths in a debianized
install of the gem. Ruby's built-in require_relative does work. The
home-grown solution causing issues was done in 1661598 to remove a
compatibility dependency. Since we're not dealing with rubies less than
1.9, we can safely ditch it.

Additionally, the C extension is installed in an arch-dependent location
in debian. We can't rely on relative paths to find it, so we just use
the regular 'require' which will know where to go.